requirements.txt
flask
numpy
Pillow
opencv-python
pytest